TOcean::ReadFrom: short loop counter for the zone-unlink walk
0x562340, 39.0% -> 47.2%. The bound test is a 16-bit `cmp word ptr [eax], bx`
(0x5623a3), so the loop variable is a short; an int counter forces a movsx of
nationCount on every compare and, downstream, blocks the strength reduction of
&contextArray[i] into the byte-offset induction the original uses.

The residual is that reduction (the original keeps a 0x48-stride offset in ESI
and re-adds the member base) plus one frame dword. Byte accounting was already
correct here and still is -- serde-audit reports the reader byte-aligned, so
this was never a desync, only shape.

+++ src/game/navy/TOcean.cpp
@@ -135,7 +135,9 @@ void TOcean::ReadFrom(TStream* stream) {
   TObject::ReadFrom(stream);
   EnsureSelectedTaskForceForOrderOwnerAndRefresh(0);
 
-  int i;
+  // A short counter: the bound test is a 16-bit `cmp word ptr [eax], bx` (0x5623a3), so
+  // an int loop variable would need a movsx of nationCount on every compare.
+  short i;
   for (i = 0; i < nationCount; ++i) {
     TZone* zone = &contextArray[i];
     if (g_pMapActionContextListHead == zone) {
@@ -215,7 +217,9 @@ void TOcean::ReadFrom(TStream* stream) {
 void TOcean::WriteTo(TStream* stream) {
   TObject::WriteTo(stream);
   stream->WriteBytes(&nationCount, 2);
-  int i;
+  // A short counter: the bound test is a 16-bit `cmp word ptr [eax], bx` (0x5623a3), so
+  // an int loop variable would need a movsx of nationCount on every compare.
+  short i;
   for (i = 0; i < nationCount; ++i) {
     contextArray[i].WriteTo(stream);
   }
